added authorization to the post and delete, and added logout controller

+++ backend/index.js
@@ -5,6 +5,7 @@ const mongoose = require('mongoose');
 const dotenv = require('dotenv');
 const usersRouter = require('./controllers/users');
 const loginRouter = require('./controllers/login');
+const logoutRouter = require('./controllers/logout');
 
 const getTokenFrom = request => {
     const authorization = request.get('authorization')
@@ -24,6 +25,9 @@ app.use(cors());
 app.use(logger);
 app.use('/api/users', usersRouter);
 app.use('/api/login', loginRouter);
+app.use('/api/logout', logoutRouter);
+
+const blacklist = [];
 
 //Get all notes
 mongoose.connect(process.env.MONGODB_CONNECTION_URL)
@@ -37,6 +41,7 @@ mongoose.connect(process.env.MONGODB_CONNECTION_URL)
 
 app.use('/users', usersRouter);
 app.use('/login', loginRouter);
+app.use('/logout', logoutRouter);
 
 const jwt = require('jsonwebtoken');
 const User = require('./models/user');
@@ -77,37 +82,39 @@ app.get('/notes/:id', async (req, res) => {
 
 // Create a new note
 app.post('/notes', async (req, res) => {
-    const { content , user , email } = req.body;
-    if (!content || !user || !email) {
-        return res.status(400).json({ error: 'Missing fields in the request' });
-    }
-
-
-    const maxId = await Note.find().sort({ id: -1 }).limit(1);
-    const newId = maxId[0].id + 1;
-
-    const token = req.headers.authorization.split(' ')[1];
-    console.log('token: ' + token);
-    const decodedToken = jwt.verify(token, process.env.SECRET)
-    if (!decodedToken.id) {
-      return response.status(401).json({ error: 'token invalid' })
-    }
-    //const user = await User.findById(decodedToken.id)
-
-    const note = new Note({
-        id: newId,
-        title: "new note",
-        author: {
-            name: user.toString(),
-            email: email.toString()
-        },
-        content: content,
-    });
-
     try {
+
+        const token = req.headers.authorization.split(' ')[1];
+        console.log('token:', token);
+        const decodedToken = jwt.verify(token, process.env.SECRET);
+        console.log('decoded token:', decodedToken);
+        if (!decodedToken.id) {
+            console.log('token invalid');
+            return response.status(401).json({ error: 'token invalid' })
+        }
+
+        const { content , user , email } = req.body;
+        if (!content || !user || !email) {
+            console.log('missing fields in request');
+            return res.status(400).json({ error: 'Missing fields in the request' });
+        }
+
+        const maxId = await Note.find().sort({ id: -1 }).limit(1);
+        const newId = maxId[0].id + 1;
+        console.log('new id', newId);
+        const note = new Note({
+            id: newId,
+            title: "new note",
+            author: {
+                name: user.toString(),
+                email: email.toString()
+            },
+            content: content,
+        });
+
         const newNote = await Note.create(note);
-        console.log('inserted note');
         res.status(200).json({note: newNote});
+
     } catch (err) {
         console.log('failed to insert note');
         res.status(500).json({ error: 'Failed to create note' });
@@ -116,40 +123,20 @@ app.post('/notes', async (req, res) => {
 
 // Update the i'th note
 app.put('/notes/:id', async (req, res) => {
-    const { content } = req.body;
-    console.log("server got: " + req.json);   
-
-    const id = parseInt(req.params.id);
-    console.log('id: ' + id);
-
-    if (!content) {
-        return res.status(400).json({ error: 'Missing fields in request' });
-    }
-    console.log(req.headers.authorization);
-
-
     try {
-        //make sure user is the author of the note
-        console.log('req.headers: ' + req);
         const token = req.headers.authorization.split(' ')[1];
-        console.log('token: ' + token);
         const decodedToken = jwt.verify(token, process.env.SECRET)
-        console.log('decodedToken: ' + decodedToken);
         if (!decodedToken.id) {
-          return response.status(403).json({ error: 'token invalid' })
+        return response.status(403).json({ error: 'token invalid' })
         }
-        else {
-            console.log('valid token');
+        const { content } = req.body;
+        const id = parseInt(req.params.id);
+        if (!content) {
+            return res.status(400).json({ error: 'Missing fields in request' });
         }
-
-
-        console.log('id: ' + id);
-        
         const note = await Note.findOneAndUpdate({ id: id }, { content : content }, { new: true });
-        console.log('note: ' + note);
-
         if (!note) 
-            return res.status(404).json({ error: 'Note not found' });
+        return res.status(404).json({ error: 'Note not found' });
         res.status(200).json({note: note});
     } catch (err) {
         res.status(500).json({ error: 'Failed to update note!' });
@@ -159,6 +146,11 @@ app.put('/notes/:id', async (req, res) => {
 // Delete the i'th note
 app.delete('/notes/:id', async (req, res) => {
     try {
+        const token = req.headers.authorization.split(' ')[1];
+        const decodedToken = jwt.verify(token, process.env.SECRET)
+        if (!decodedToken.id) {
+          return response.status(403).json({ error: 'token invalid' })
+        }
         const note = await Note.findOneAndDelete({ id: req.params.id });
         if (!note) 
             return res.status(404).json({ error: 'Note not found' });
@@ -167,6 +159,7 @@ app.delete('/notes/:id', async (req, res) => {
         res.status(500).json({ error: 'Failed to delete note' });
     }
 });
+
 
 app.listen(PORT, () => {
     console.log(`Server running on port ${PORT}`);
